// src/commons/models/productPromotion.model.ts
export enum CodeEnum {
  EMBLEM_ELITE = "emblem_elite",
  TOKEN_30000 = "token_30000",
  TOKEN_13500 = "token_13500",
  TOKEN_5000 = "token_5000",
  TOKEN_2000 = "token_2000",
  TOKEN_1000 = "token_1000",
  TOKEN_500 = "token_500",
}

export enum DiscountTypeEnum {
  PERCENT_OFF = "PERCENT_OFF",
  AMOUNT_OFF = "AMOUNT_OFF",
}